match func_802C3784_645C34 by splitting the file
