+++ splat.yaml
@@ -3667,10 +3667,12 @@ segments:
     - [0x6438F0, c]
     - [0x644E80, c]
     - [0x645740, c]
+    - [0x645870, c]
     - [0x645D90, c]
     - [0x646B50, c]
     - [0x646EA0, c]
     - [0x647360, c]
+
     - [0x648600, bin] # , data]
 
     - [0x64A500, .rodata, 6401B0]
